Moved GOCoupler and GODivisionalCoupler and GODrawStop to the model subdirectory
