+++ src/grandorgue/GOOrganController.cpp
@@ -48,6 +48,8 @@
 #include "midi/GOMidiEvent.h"
 #include "midi/GOMidiPlayer.h"
 #include "midi/GOMidiRecorder.h"
+#include "model/GOCoupler.h"
+#include "model/GODivisionalCoupler.h"
 #include "model/GOEnclosure.h"
 #include "model/GOManual.h"
 #include "model/GORank.h"
@@ -65,8 +67,6 @@
 #include "GOBuffer.h"
 #include "GOCache.h"
 #include "GOCacheWriter.h"
-#include "GOCoupler.h"
-#include "GODivisionalCoupler.h"
 #include "GODocument.h"
 #include "GOEvent.h"
 #include "GOHash.h"
