+++ src/grandorgue/combinations/model/GODivisionalCombination.cpp
@@ -15,9 +15,9 @@
 #include "combinations/GOSetter.h"
 #include "config/GOConfigReader.h"
 #include "config/GOConfigWriter.h"
+#include "model/GODivisionalCoupler.h"
 #include "model/GOManual.h"
 
-#include "GODivisionalCoupler.h"
 #include "GOOrganController.h"
 
 GODivisionalCombination::GODivisionalCombination(
