+++ src/grandorgue/gui/GOGUIPanel.cpp
@@ -15,14 +15,14 @@
 #include "config/GOConfigReader.h"
 #include "config/GOConfigWriter.h"
 #include "control/GOPistonControl.h"
+#include "model/GOCoupler.h"
+#include "model/GODivisionalCoupler.h"
 #include "model/GOManual.h"
 #include "model/GOStop.h"
 #include "model/GOSwitch.h"
 #include "model/GOTremulant.h"
 
-#include "GOCoupler.h"
 #include "GODC.h"
-#include "GODivisionalCoupler.h"
 #include "GOGUIButton.h"
 #include "GOGUIControl.h"
 #include "GOGUIDisplayMetrics.h"
