+++ src/grandorgue/model/GODivisionalCoupler.h
